Smile-Classifier/requirements.txt
numpy
pandas
matplotlib
seaborn
opencv-python
scikit-learn
tensorflow
python-dotenv
fastapi
uvicorn
sqlalchemy
pillow
pymysql
python-multipart
jinja2
cryptography